we add a new hook
new hook usefacelocation
