+++ src/App.js
@@ -2,6 +2,7 @@
 import './App.css';
 import React, { useEffect, useState } from 'react';
 import useFetch from "./Hooks/useFetch.js";
+import useFacelocation from './Hooks/useFacelocation';
 import Navigator from "./Components/Navigator/Navidator";
 import Particles from "react-tsparticles";
 import { loadFull } from "tsparticles";
@@ -61,20 +62,20 @@ function App(){
   
 
   //CALCULATE FACE LOCATION(BOX) ////////////
-  const FaceLocation = (data) => {
-    const faceDataLocation = data.outputs[0].data.regions[0].region_info.bounding_box;
-    const image = document.getElementById("ImageInput");
-    const img_width = Number(image.width);
-    const img_height = Number(image.height);
-    console.log(img_width, img_height, faceDataLocation)
+  // const FaceLocation = async (data) => {
+  //   const faceDataLocation = data.outputs[0].data.regions[0].region_info.bounding_box;
+  //   const image = document.getElementById("ImageInput");
+  //   const img_width = Number(image.width);
+  //   const img_height = Number(image.height);
+  //   console.log(img_width, img_height, faceDataLocation)
 
-    return {
-      left_Column: (faceDataLocation.left_col * img_width),
-      right_Column: (img_width - (faceDataLocation.right_col * img_width)),
-      bottom_row: (img_height - (faceDataLocation.bottom_row * img_height)),
-      top_row: (faceDataLocation.top_row * img_height),
-    }
-  }
+  //   return {
+  //     left_Column: (faceDataLocation.left_col * img_width),
+  //     right_Column: (img_width - (faceDataLocation.right_col * img_width)),
+  //     bottom_row: (img_height - (faceDataLocation.bottom_row * img_height)),
+  //     top_row: (faceDataLocation.top_row * img_height),
+  //   }
+  // }
 
   ////////////////INPUT ////////////
   const [input, setInput] = useState(""); 
@@ -86,17 +87,19 @@ function App(){
 
   /////////BUTTON PRINCIPAL PAGE///////////////////
   const [imageUrl,setImageUrl]=useState("");
-  const [data,loading, error] = useFetch(imageUrl);
+  const {data,refetch} = useFetch(imageUrl);
+  const {box:boxlocation} = useFacelocation(data);
    
   const onButtonChange = async () => {
+    await imageset()
+    console.log("data: " + data)
+    displayFacebox(boxlocation);  
     
-     setImageUrl(input);
-    if (data!==null){
-       displayFacebox(FaceLocation(data))
-    }  
   }
- 
-    
+  const imageset = async () =>{
+    setImageUrl(input);
+
+  }
   //    fetch("https://api.clarifai.com/v2/models/" + MODEL_ID + "/versions/" + MODEL_VERSION_ID + "/outputs", requestOptions)
   //     .then(response => response.json())
   //     .then(result =>displayFacebox(FaceLocation(result)))
